+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### HEAD
 
+* Add configs for web application manifest files
+  [[#29](https://github.com/h5bp/server-configs-apache/issues/29)].
 * Don't block access to the content from within the `/.well-known/` directory
   [[#31](https://github.com/h5bp/server-configs-apache/issues/31)].
 * Forbid access to `.conf` files.